Apple Health façade integrated into factory alongside Healthmate Fixes #18
